make it in tsx and ts
